requirements.txt
jinja2 ~= 3.1.3
markdown-it-py[plugins] ~= 3.0.0
mdit-py-figure @ git+https://github.com/mpaulon/mdit-py-figure
requests ~= 2.32.3